separate scaladsl/javadsl and avoid overloaded spawn, #24470
* copy and amend scaladoc for ClusterSharding extension
