+++ akka-cluster-sharding-typed/src/main/protobuf/ShardingMessages.proto
@@ -9,5 +9,5 @@ import "ContainerFormats.proto";
 
 message ShardingEnvelope {
   required string entityId = 1;
-  optional Payload message = 2;
+  required Payload message = 2;
 }
